requirements.txt
# Core dependencies
gradio>=4.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# MCP Server dependencies
fastmcp>=0.2.0
uvicorn>=0.27.0
pydantic>=2.0.0

# Research paper tools
arxiv>=2.0.0
pypdf>=3.0.0

# API clients for source verification
requests>=2.31.0

# Standard library dependencies (included for completeness, but not needed to install)
# json, logging, os, datetime, time, re, typing are part of Python's standard library
